requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

Babel>=1.3
pbr>=1.6
cliff>=1.14.0  # Apache-2.0
argparse
iso8601>=0.1.9
PrettyTable<0.8,>=0.7
oslo.i18n>=1.5.0 # Apache-2.0
oslo.serialization>=1.4.0 # Apache-2.0
oslo.utils>=2.4.0 # Apache-2.0
python-keystoneclient!=1.8.0,>=1.6.0
python-swiftclient>=2.2.0
python-openstackclient>=1.5.0
PyYAML>=3.1.0
requests!=2.8.0,>=2.5.2
six>=1.9.0
